requirements.txt
django-mssql-backend==2.8.1
mssql-django==1.5
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
CacheControl==0.13.1
cachetools==5.3.2
click>=8.1.2,<9.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==2.0.12
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==2.1
django-multiupload==0.6.1
django-redis==5.4.0
django-restframework==0.0.1
django-webpush==0.3.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
flower==2.0.1
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.0.4
h2==4.1.0
hpack==4.0.0
http-ece==1.2.0
httplib2==0.22.0
humanize==4.9.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isodate==0.6.1
Jinja2==2.11.2
kombu==5.3.5
libretranslatepy==2.1.1
lxml==5.0.0
MarkupSafe==2.1.3
msgpack==1.0.7
mysqlclient==2.2.1
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==4.1.0
priority==1.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
py-vapid==1.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywebpush==1.9.4
redis==5.0.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tornado==6.4
translate==3.6.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
wfastcgi==3.0.0
whitenoise==6.6.0
XlsxWriter==3.1.9
zeep==4.2.1
zope.interface==6.1